app/requirements.txt
asgiref==3.3.4
Django==3.1.8
django-djconfig==0.10.0
django-environ==0.4.5
django-haystack==3.0
django-health-check==3.16.4
django-infinite-scroll-pagination==1.1.0
django-spirit==0.12.2
gunicorn==20.1.0
mistune==0.8.4
olefile==0.46
Pillow==8.0.1
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
Whoosh==2.7.4
